+++ core/modules/user/hooks/beforeRegistration.ts
@@ -7,13 +7,13 @@ export function beforeRegistration({ Vue, config, store, isServer }) {
   const dbNamePrefix = storeView.storeCode ? storeView.storeCode + '-' : ''
 
   Vue.prototype.$db.usersCollection = new UniversalStorage(localForage.createInstance({
-    name: (config.cart.multisiteCommonCart ? '' : dbNamePrefix) + 'shop',
+    name: (config.storeViews.commonCache ? '' : dbNamePrefix) + 'shop',
     storeName: 'user',
     driver: localForage[config.localForage.defaultDrivers['user']]
   }))
 
   Vue.prototype.$db.ordersHistoryCollection = new UniversalStorage(localForage.createInstance({
-    name: (config.cart.multisiteCommonCart ? '' : dbNamePrefix) + 'shop',
+    name: (config.storeViews.commonCache ? '' : dbNamePrefix) + 'shop',
     storeName: 'ordersHistory',
     driver: localForage[config.localForage.defaultDrivers['ordersHistory']]
   }))